Using ssh keys to generate a secret that is used to encrypt the data
Many thanks for the discussion in realestate-com-au/credulous#60

As part of this I also made the signature in the credentials.json just
of the secret rather than also include information about the data. This
should make it easier to make credo support more than just amazon creds
in the future.