Adding serve and switch commands.
The serve command creates a fake metadata service on 169.254.169.254:80

The switch command tells the server which credentials to be serving.

The server will cache the basic auth for ldap in memory for four hours
before requiring them to be provided again.

The server will also cache the credentials in memory and redo the assume
role if they have expired.